+++ tests/test_models.py
@@ -30,6 +30,9 @@ def test_device(attributes: HvacAttributes):
     assert device.get_device_code() == attributes.device_code
     assert device.get_model() == attributes.model
     assert device.get_name() == attributes.name
+    assert device.get_serial_number() == attributes.serial_number
+    assert device.get_manufacturer() == attributes.manufacturer
+    assert device.get_firmware_version() == attributes.firmware_version
     assert device.get_system_mode() == attributes.system_mode
     assert device.get_running_state() == attributes.running_state
     assert device.get_target_temperature() == attributes.target_temperature